Fixing int to float in docs for sigint_timeout and sigterm_timeout

+++ tools/roslaunch/src/roslaunch/child.py
@@ -68,10 +68,10 @@ class ROSLaunchChild(object):
         @param server_uri: XML-RPC URI of roslaunch server
         @type  server_uri: str
         @param sigint_timeout: The SIGINT timeout used when killing nodes (in seconds).
-        @type  sigint_timeout: int
+        @type  sigint_timeout: float
         @param sigterm_timeout: The SIGTERM timeout used when killing nodes if SIGINT does not stop the node (
                                 in seconds).
-        @type  sigterm_timeout: int
+        @type  sigterm_timeout: float
         @return: XML-RPC URI
         @rtype:  str
         """
